Update organization.md
I need a review for the group organization. If anyone has a preference, please let me know, like I do not have to be the team lead. If someone else wants that title, I am fine with it. If Jack does not want to be the tester, he can switch it up with me or someone else.